requirements.txt
flask-bootstrap
Bootstrap_Flask
Flask_Login
Flask_WTF
WTForms
Werkzeug
Flask
flask_sqlalchemy
SQLAlchemy
langchain
langchain-core
langchain-community
langchain-openai
pydantic
duckduckgo-search
wikipedia
python-dotenv
pytest
gunicorn
psycopg2-binary